исправлен обход дерева технологий. Теперь обходим по одному и в кажом дереве ищем проблемы.
